fix loop checking for redirects in urls
Before this, the try code was always causing an exception because it was using requests directly
(And requests was not imported).
But even with importing requests, an error is thrown because it should be using an async method.
This refactors it to use the async_get_clientsession(), and now it will correctly follow redirects.